backend/requirements.txt
flask
flask_cors
flask-restx
flask-sqlalchemy
Flask-Migrate
sqlalchemy
alembic
gunicorn
mysqlclient
pymysql
unittest-xml-reporting
requests
cryptography
python-dotenv
dependency-injector
SQLAlchemy-serializer
pydantic[all]
email-validator
pyjwt
lcax==2.6.3